refactor identity bounded context
