+++ services/backend/src/swen_identity/application/commands/create_user_command.py
@@ -3,13 +3,13 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from swen_identity.application.ports.unit_of_work import UnitOfWork
-from swen_identity.domain.user import (
+from swen_identity.domain import (
     EmailAlreadyExistsError,
     User,
+    UserCredentialRepository,
     UserRepository,
     UserRole,
 )
-from swen_identity.repositories import UserCredentialRepository
 from swen_identity.services import PasswordHashingService
 
 if TYPE_CHECKING:
