+++ services/backend/src/swen_identity/domain/aggregates/__init__.py
@@ -1,0 +1,9 @@
+"""User aggregates."""
+
+from swen_identity.domain.aggregates.password_reset_token import (
+    PasswordResetToken,
+)
+from swen_identity.domain.aggregates.user import User
+from swen_identity.domain.aggregates.user_credential import UserCredential
+
+__all__ = ["PasswordResetToken", "User", "UserCredential"]
