+++ services/backend/src/swen_identity/domain/exceptions.py
