+++ services/backend/src/swen_identity/domain/repositories/__init__.py
@@ -1,0 +1,15 @@
+"""User repositories."""
+
+from swen_identity.domain.repositories.password_reset_token_repository import (
+    PasswordResetTokenRepository,
+)
+from swen_identity.domain.repositories.user_credential_repository import (
+    UserCredentialRepository,
+)
+from swen_identity.domain.repositories.user_repository import UserRepository
+
+__all__ = [
+    "PasswordResetTokenRepository",
+    "UserCredentialRepository",
+    "UserRepository",
+]
