--- services/backend/src/swen_identity/domain/user/aggregates/__init__.py
@@ -1,5 +1,0 @@
-"""User aggregates."""
-
-from swen_identity.domain.user.aggregates.user import User
-
-__all__ = ["User"]
