--- services/backend/src/swen_identity/domain/user/repositories/__init__.py
@@ -1,5 +1,0 @@
-"""User repositories."""
-
-from swen_identity.domain.user.repositories.user_repository import UserRepository
-
-__all__ = ["UserRepository"]
