+++ services/backend/src/swen_identity/domain/value_objects/__init__.py
@@ -1,0 +1,11 @@
+"""Value objects for the user domain having identity concerns only."""
+
+from swen_identity.domain.value_objects.email import Email
+from swen_identity.domain.value_objects.token_payload import TokenPayload
+from swen_identity.domain.value_objects.user_role import UserRole
+
+__all__ = [
+    "Email",
+    "TokenPayload",
+    "UserRole",
+]
