+++ services/backend/src/swen_identity/infrastructure/persistence/sqlalchemy/repositories/user_repository.py
@@ -8,7 +8,7 @@ from sqlalchemy import func, select
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from swen_identity.domain.user import (
+from swen_identity.domain import (
     Email,
     EmailAlreadyExistsError,
     User,
